extract shorthand properties to a trait
